# chore: Enable quality gate in smoke e2e pipelines for account listing