--- a/e2e/specs/performance/account-list/render-account-list.spec.ts
+++ b/e2e/specs/performance/account-list/render-account-list.spec.ts
@@ -31,10 +31,10 @@ describe(SmokePerformance('Account List Load Testing'), () => {
         const isAndroid = device.getPlatform() === 'android';
         const PERFORMANCE_THRESHOLDS = isAndroid
           ? {
-              TOTAL_TIME: 5000, // 5 seconds max for Android
+              TOTAL_TIME: 5900, // 5.9 seconds max for Android
             }
           : {
-              TOTAL_TIME: 5000, // 5 seconds max for iOS
+              TOTAL_TIME: 4000, // 4 seconds max for iOS
             };
 
         console.log(
@@ -81,7 +81,7 @@ describe(SmokePerformance('Account List Load Testing'), () => {
 
             // Performance assertions with warnings
             if (totalTime > PERFORMANCE_THRESHOLDS.TOTAL_TIME) {
-              console.warn(
+              throw new Error(
                 `Performance test failed: Total time (${totalTime}ms) exceeded maximum acceptable time (${PERFORMANCE_THRESHOLDS.TOTAL_TIME}ms)`,
               );
             }
@@ -124,10 +124,10 @@ describe(SmokePerformance('Account List Load Testing'), () => {
         const isAndroid = device.getPlatform() === 'android';
         const HEAVY_LOAD_THRESHOLDS = isAndroid
           ? {
-              TOTAL_TIME: 5000, // 5 seconds max for Android
+              TOTAL_TIME: 4200, // 4.2 seconds max for Android
             }
           : {
-              TOTAL_TIME: 5000, // 5 seconds max for iOS
+              TOTAL_TIME: 4200, // 4.2 seconds max for iOS
             };
 
         let result: Partial<TestResult> = {};
@@ -165,7 +165,7 @@ describe(SmokePerformance('Account List Load Testing'), () => {
             );
 
             if (totalTime > HEAVY_LOAD_THRESHOLDS.TOTAL_TIME) {
-              console.warn(
+              throw new Error(
                 `Heavy load test failed: Total time (${totalTime}ms) exceeded maximum acceptable time (${HEAVY_LOAD_THRESHOLDS.TOTAL_TIME}ms)`,
               );
             }
@@ -193,10 +193,10 @@ describe(SmokePerformance('Account List Load Testing'), () => {
         const isAndroid = device.getPlatform() === 'android';
         const BASELINE_THRESHOLDS = isAndroid
           ? {
-              TOTAL_TIME: 5000, // 5 seconds max for Android
+              TOTAL_TIME: 3800, // 3.8 seconds max for Android
             }
           : {
-              TOTAL_TIME: 5000, // 5 seconds max for iOS
+              TOTAL_TIME: 3800, // 3.8 seconds max for iOS
             };
         // Baseline test with minimal tokens for comparison
         const minimalTokens = [
@@ -250,7 +250,7 @@ describe(SmokePerformance('Account List Load Testing'), () => {
 
             // Baseline should be very fast
             if (totalTime > BASELINE_THRESHOLDS.TOTAL_TIME) {
-              console.warn(
+              throw new Error(
                 `Baseline test failed: Total time (${totalTime}ms) exceeded maximum acceptable time (${BASELINE_THRESHOLDS.TOTAL_TIME}ms)`,
               );
             }