<!--
Please submit this PR as a draft initially.
Do not mark it as "Ready for review" until the template has been completely filled out, and PR status checks have passed at least once.
-->

## **Description**
Enable quality gate in smoke e2e pipelines for account listing

We've ran 10 bitrise pipelines to get minum, max, average and delta results from tests running
* https://app.bitrise.io/app/be69d4368ee7e86d/pipelines/a0554c31-25d5-4a68-bcbe-7bcb4c107263?tab=artifacts
* https://app.bitrise.io/app/be69d4368ee7e86d/pipelines/747045b1-58a6-4986-9b99-982aa915963e?tab=artifacts
* https://app.bitrise.io/app/be69d4368ee7e86d/pipelines/6c881b29-1f71-4d60-8a57-c5d2db40c448?tab=artifacts
* https://app.bitrise.io/app/be69d4368ee7e86d/pipelines/34ae3782-6624-45dc-9dbc-432a5aaab4b4?tab=artifacts
* https://app.bitrise.io/app/be69d4368ee7e86d/pipelines/d3a240c1-31f4-49e2-9d43-ca18c29df034?tab=artifacts
* https://app.bitrise.io/app/be69d4368ee7e86d/pipelines/ddf54e8a-17cc-4f38-8d7d-8d4d07caa110?tab=artifacts
* https://app.bitrise.io/app/be69d4368ee7e86d/pipelines/6f54ac83-9da5-473e-8edc-2ead1619b167?tab=artifacts
* https://app.bitrise.io/app/be69d4368ee7e86d/pipelines/ef411c36-2de0-4e1f-ad41-c409d01c4eac?tab=artifacts
* https://app.bitrise.io/app/be69d4368ee7e86d/pipelines/347accd0-47c4-4316-9d2d-de1ea3f81d6e?tab=artifacts
* https://app.bitrise.io/app/be69d4368ee7e86d/pipelines/c5dc140c-4c09-4087-840c-51f8cfae7426?tab=artifacts

Results:
<img width="1091" height="362" alt="image" src="https://github.com/user-attachments/assets/8bd9ae85-0aa6-4e4b-9221-fdf00adc3cf6" />
<img width="1076" height="358" alt="image" src="https://github.com/user-attachments/assets/7712450a-fea9-491e-8d8b-35e7d352850c" />


New quality gates thresholds proposed:
## 🏠 Account List Performance Tests

### Test: `render account list efficiently with multiple accounts and networks`

**Configuration**: Multiple accounts, popular networks, profile syncing enabled

| Platform    | Total Max Time | Notes               |
| ----------- | -------------- | ------------------- |
| **Android** | 5900ms         | 5.9 seconds maximum |
| **iOS**     | 4000ms         | 4 seconds maximum   |

---

### Test: `handle account list performance with heavy token load`

**Configuration**: Multiple accounts, popular networks, 10 tokens for stress testing

| Platform    | Total Max Time | Notes               |
| ----------- | -------------- | ------------------- |
| **Android** | 4200ms         | 4.2 seconds maximum |
| **iOS**     | 4200ms         | 4.2 seconds maximum |

---

### Test: `benchmark account list with minimal load`

**Configuration**: Minimal accounts, default network, 2 tokens (baseline measurement)

| Platform    | Total Max Time | Notes               |
| ----------- | -------------- | ------------------- |
| **Android** | 3800ms         | 3.8 seconds maximum |
| **iOS**     | 3800ms         | 3.8 seconds maximum |

---


Test on purpose failure of the performance regression pipeline testing (still running):
https://app.bitrise.io/app/be69d4368ee7e86d/pipelines/574f564e-5430-49c1-a6c6-3d10da617740



<!--
Write a short description of the changes included in this pull request, also include relevant motivation and context. Have in mind the following questions:
1. What is the reason for the change?
2. What is the improvement/solution?
-->

## **Changelog**

<!--
If this PR is not End-User-Facing and should not show up in the CHANGELOG, you can choose to either:
1. Write `CHANGELOG entry: null`
2. Label with `no-changelog`

If this PR is End-User-Facing, please write a short User-Facing description in the past tense like:
`CHANGELOG entry: Added a new tab for users to see their NFTs`
`CHANGELOG entry: Fixed a bug that was causing some NFTs to flicker`

(This helps the Release Engineer do their job more quickly and accurately)
-->

CHANGELOG entry:

## **Related issues**

Fixes:

## **Manual testing steps**

1. Go to this page...
2.
3.

## **Screenshots/Recordings**

<!-- If applicable, add screenshots and/or recordings to visualize the before and after of your change. -->

### **Before**

<!-- [screenshots/recordings] -->

### **After**

<!-- [screenshots/recordings] -->

## **Pre-merge author checklist**

- [ ] I’ve followed [MetaMask Contributor Docs](https://github.com/MetaMask/contributor-docs) and [MetaMask Mobile Coding Standards](https://github.com/MetaMask/metamask-mobile/blob/main/.github/guidelines/CODING_GUIDELINES.md).
- [ ] I've completed the PR template to the best of my ability
- [ ] I’ve included tests if applicable
- [ ] I’ve documented my code using [JSDoc](https://jsdoc.app/) format if applicable
- [ ] I’ve applied the right labels on the PR (see [labeling guidelines](https://github.com/MetaMask/metamask-mobile/blob/main/.github/guidelines/LABELING_GUIDELINES.md)). Not required for external contributors.

## **Pre-merge reviewer checklist**

- [ ] I've manually tested the PR (e.g. pull and build branch, run the app, test code being changed).
- [ ] I confirm that this PR addresses all acceptance criteria described in the ticket it closes and includes the necessary testing evidence such as recordings and or screenshots.